Refactor generateDiff method to add JSON validation
I didn't realize at first that the JSON.parse() method will fail handily and can break my comparison in a bunch of ways, so I added a validation method to ensure that it doesn't happen.

If, for example, one of the shop listing files was left empty but still existed, JSON.parse() would fail, and the diff would fail, and we'd get a whole bunch of errors.